Move DateRange class into its own file
