+++ Template_Skill/src/main/java/com/neong/voice/wolfpack/CalendarConversation.java
@@ -6,13 +6,9 @@ import com.amazon.speech.speechlet.IntentRequest;
 import com.amazon.speech.speechlet.Session;
 import com.amazon.speech.speechlet.SpeechletResponse;
 
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
 import com.neong.voice.wolfpack.CalendarHelper;
 import com.neong.voice.wolfpack.CosineSim;
+import com.neong.voice.wolfpack.DateRange;
 
 import com.neong.voice.model.base.Conversation;
 
@@ -587,44 +583,5 @@ public class CalendarConversation extends Conversation {
 	 */
 	private SpeechletResponse newBadStateResponse() {
 		return newTellResponse("Sorry, I forgot what we were talking about.", false);
-	}
-}
-
-
-@JsonAutoDetect(fieldVisibility=Visibility.ANY,
-                getterVisibility=Visibility.NONE,
-                isGetterVisibility=Visibility.NONE)
-class DateRange {
-	// The values stored in session attributes must be convertable to JSON.
-	// java.sql.Date does not have this ability, so we use a String instead.
-	@JsonProperty("date") private final Date _date;
-	@JsonProperty("range") private final String _range;
-
-	public DateRange(String dateString) {
-		// TODO: actual implementation handling weeks, months, etc.
-		_date = Date.valueOf(dateString);
-		_range = "day";
-	}
-
-	@JsonCreator
-	public DateRange(Map<String, Object> props) {
-		_date = Date.valueOf((String) props.get("date"));
-		_range = (String) props.get("range");
-	}
-
-	public Date getDate() {
-		return _date;
-	}
-
-	public Timestamp getTimestamp() {
-		return Timestamp.valueOf(_date.toString() + " 00:00:00");
-	}
-
-	public String getRange() {
-		return _range;
-	}
-
-	public String getDateSsml() {
-		return CalendarHelper.formatDateSsml(getTimestamp());
 	}
 }
